+++ src/main/java/com/example/askme/api/controller/login/OAuthController.java
@@ -6,8 +6,8 @@ import com.example.askme.api.controller.login.validator.OauthValidator;
 import com.example.askme.api.service.oauth.OauthLoginService;
 import com.example.askme.common.ResultResponse;
 import com.example.askme.common.constant.LoginType;
-import com.example.askme.common.resolver.memberInfo.TokenDto;
-import com.example.askme.common.resolver.memberInfo.TokenParser;
+import com.example.askme.common.resolver.tokenInfo.TokenDto;
+import com.example.askme.common.resolver.tokenInfo.TokenParser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
