+++ src/main/java/com/example/askme/api/controller/token/controller/TokenController.java
@@ -3,8 +3,8 @@ package com.example.askme.api.controller.token.controller;
 import com.example.askme.api.controller.token.dto.AccessTokenResponseDto;
 import com.example.askme.common.ResultResponse;
 import com.example.askme.common.jwt.TokenManager;
-import com.example.askme.common.resolver.memberInfo.TokenDto;
-import com.example.askme.common.resolver.memberInfo.TokenParser;
+import com.example.askme.common.resolver.tokenInfo.TokenDto;
+import com.example.askme.common.resolver.tokenInfo.TokenParser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
