+++ src/main/java/com/example/askme/common/resolver/tokenInfo/TokenDto.java
@@ -1,4 +1,4 @@
-package com.example.askme.common.resolver.memberInfo;
+package com.example.askme.common.resolver.tokenInfo;
 
 import lombok.Builder;
 import lombok.Getter;
